requirements.txt
Flask>=2.3
SQLAlchemy>=2.0
Flask-SQLAlchemy>=3.1
APScheduler>=3.10
python-dateutil>=2.8
pydantic>=2.7
flask-sock>=0.7.0
huggingface_hub>=0.23
regex>=2024.5.15
python-dotenv>=1.0
streamlit>=1.34
